refactor(file): R2관련 로직을 따로 도메인 폴더로 구분
